Extract bitmap and set commands
