+++ fakeredis/commands_mixins/bitmap_mixin.py
@@ -1,0 +1,58 @@
+import random
+from fakeredis import _msgs as msgs
+from fakeredis._commands import (command, Key, Int, Float, CommandItem, BitOffset, BitValue)
+from fakeredis._helpers import (OK, SimpleError, casematch, MAX_STRING_SIZE, fix_range_string)
+
+
+class BitmapCommandsMixin:
+    # Key commands
+    # TODO: lots
+
+    @command((Key(bytes, 0),), (bytes,))
+    def bitcount(self, key, *args):
+        # Redis checks the argument count before decoding integers. That's why
+        # we can't declare them as Int.
+        if args:
+            if len(args) != 2:
+                raise SimpleError(msgs.SYNTAX_ERROR_MSG)
+            start = Int.decode(args[0])
+            end = Int.decode(args[1])
+            start, end = fix_range_string(start, end, len(key.value))
+            value = key.value[start:end]
+        else:
+            value = key.value
+        return bin(int.from_bytes(value, 'little')).count('1')
+
+    @command((Key(bytes), BitOffset))
+    def getbit(self, key, offset):
+        value = key.get(b'')
+        byte = offset // 8
+        remaining = offset % 8
+        actual_bitoffset = 7 - remaining
+        try:
+            actual_val = value[byte]
+        except IndexError:
+            return 0
+        return 1 if (1 << actual_bitoffset) & actual_val else 0
+
+    @command((Key(bytes), BitOffset, BitValue))
+    def setbit(self, key, offset, value):
+        val = key.get(b'\x00')
+        byte = offset // 8
+        remaining = offset % 8
+        actual_bitoffset = 7 - remaining
+        if len(val) - 1 < byte:
+            # We need to expand val so that we can set the appropriate
+            # bit.
+            needed = byte - (len(val) - 1)
+            val += b'\x00' * needed
+        old_byte = val[byte]
+        if value == 1:
+            new_byte = old_byte | (1 << actual_bitoffset)
+        else:
+            new_byte = old_byte & ~(1 << actual_bitoffset)
+        old_value = value if old_byte == new_byte else 1 - value
+        reconstructed = bytearray(val)
+        reconstructed[byte] = new_byte
+        key.update(bytes(reconstructed))
+        return old_value
